feat: rename reducer to cashReducer; create customerReducer reducer; decompose reducers into separate files; unite reducers with combineReducers() method; change refs to cash in CashCounter
