+++ src/Store/types/reducer.types.ts
@@ -1,0 +1,13 @@
+export type ActionType = {
+  type: string;
+  payload: number;
+};
+
+export type CashDefaultStateType = {
+  cash: number;
+};
+
+export type CustomerDefaultStateType = {
+  customers: string[]
+};
+
